Add files via upload
Greeting, Students!

This HTML code defines a webpage with a script that will run when the page loads (‘onload’ event). The function ‘showMessage’ will execute an alert and also print a message to the browser's console. To see the console message, users will need to open the developer tools in their browser (usually F12 or right-click and select "Inspect", then click on the "Console" tab).

Happy a Learning Day!
https://www.linkedin.com/in/farshid-keivanian
https://keivan-rc.com/